support vw/vh units for font size
